Central file for constant values.
UPDATED: UPDATED: The client's increase_nitrate method now expects a parameter specifying the amount of nitrate to be added to the nitrate pool.

UPDATED: All constants of Plant, Water, Starch, and Nitrate were moved to constant.py. Here all constants are collected in a central place with their references. Furthermore, constant.py contains some customizable parameters that influence, for example, the usage of the pools by the plant.

UPDATED: There are new test cases for dynamic model and plant and the tests for starch have been adjusted.

+++ src/PlantEd/client/growth_percentage.py
@@ -25,7 +25,7 @@ class GrowthPercent:
     root: float
     starch: float
     flower: float
-    time_frame: float
+    time_frame: int
 
     @classmethod
     def from_json(self, string:str) -> GrowthPercent:
